# Farmease-BE/farmease/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/domba": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all sheep with filtering by pen, gender, and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dombas"
                ],
                "summary": "Get list of all sheep",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by pen ID",
                        "name": "id_kandang",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by gender (Jantan/Betina)",
                        "name": "jenis_kelamin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by tag or nickname",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Domba"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new sheep entry with initial details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dombas"
                ],
                "summary": "Register a new sheep",
                "parameters": [
                    {
                        "description": "Sheep details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Domba"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Domba"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            }
        },
...
        "/api/domba/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve specific sheep details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dombas"
                ],
                "summary": "Get details of a sheep",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sheep ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Domba"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            }
        },
...
        "/api/domba/{id}/silsilah": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve family tree of a sheep up to specified generations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dombas"
                ],
                "summary": "Get sheep genealogy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sheep ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of generations to traverse",
                        "name": "generasi",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Silsilah"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            }
        },
        "/api/domba/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the status (e.g., Sehat, Sakit, Terjual) of a sheep",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dombas"
                ],
                "summary": "Update sheep status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sheep ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all users with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get list of users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by external subject",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-array_http_GetUserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Synchronize user data with external identity provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Sync user data",
                "parameters": [
                    {
                        "description": "Sync details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SyncUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-http_SyncUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Domba": {
            "type": "object",
            "properties": {
                "asal": {
                    "type": "string"
                },
                "berat_terakhir": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id_domba": {
                    "type": "integer"
                },
                "id_induk_betina": {
                    "type": "integer"
                },
                "id_induk_jantan": {
                    "type": "integer"
                },
                "id_jenis": {
                    "type": "integer"
                },
                "id_kandang": {
                    "type": "integer"
                },
                "induk_betina": {
                    "$ref": "#/definitions/domain.Parent"
                },
                "induk_jantan": {
                    "$ref": "#/definitions/domain.Parent"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "kode_domba": {
                    "type": "string"
                },
                "nama_domba": {
                    "type": "string"
                },
                "nama_jenis": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Parent": {
            "type": "object",
            "properties": {
                "id_domba": {
                    "type": "integer"
                },
                "nama_domba": {
                    "type": "string"
                }
            }
        },
...
        "domain.Silsilah": {
            "type": "object",
            "properties": {
                "generasi": {
                    "type": "integer"
                },
                "id_domba": {
                    "type": "integer"
                },
                "induk_betina": {
                    "$ref": "#/definitions/domain.Silsilah"
                },
                "induk_jantan": {
                    "$ref": "#/definitions/domain.Silsilah"
                },
                "nama_domba": {
                    "type": "string"
                }
            }
        },
...
        "http.GetUserResponse": {
            "type": "object",
            "properties": {
                "external_subject": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "institution_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "http.SyncUserRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "http.SyncUserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "responses.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "responses.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/responses.Error"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "responses.Response-array_http_GetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.GetUserResponse"
                    }
                },
                "error": {
                    "$ref": "#/definitions/responses.Error"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "responses.Response-http_SyncUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/http.SyncUserResponse"
                },
                "error": {
                    "$ref": "#/definitions/responses.Error"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}